==> 00-3. InpAnlyzSevPaybase_All.py <==
'''
This script is written 4 anyalyzing All's Inpatients' Data based on 'DRGNO by the order of payment'.

Written Date: 2019.12.17.
Written By: Peter JH Park

'''

### Import modules in needs

import os, sys, csv
import pandas as pd
from pandas import DataFrame
import numpy as np
import matplotlib.pyplot as plt
import datetime, time
import re
from itertools import product
import math

print("\n Current Working Directory is: ", os.getcwd())

### READ Files & Check

### READ Files & Check

SNUCHIn = pd.read_csv("./SNUCH/SNUCHInP_R4A.csv", encoding="utf-8")
#KNUCHIn = pd.read_csv("./KNUCH/KNUCHInP_R4A.csv", encoding="utf-8")
#SACHIn = pd.read_csv("./SACH/SACHInP_R4A.csv", encoding="utf-8")
#SCHIn = pd.read_csv("./SCH/SCHInP_R4A.csv", encoding="utf-8")
PNUCHIn = pd.read_csv("./PNUCH/PNUCHInP_R4A.csv", encoding="utf-8")
JNNUCHIn = pd.read_csv("./JNNUCH/JNNUCHInP_R4A.csv", encoding="utf-8")
JBNUCHIn = pd.read_csv("./JBNUCH/JBNUCHInP_R4A.csv", encoding="utf-8")
KBNUCHIn = pd.read_csv("./KBNUCH/KBNUCHInP_R4A.csv", encoding="utf-8")

DiagCN = pd.read_csv("./master_Dcode&Dname.csv", encoding="utf-8")
SurCN = pd.read_csv("./master_Scode&Sname.csv", encoding="utf-8")
DRGCNmain5 = pd.read_csv("./SevCri_5digit(DRGname).csv", encoding="utf-8")
DRGCNsub4 = pd.read_csv("./SevCri_4digit(DRGname).csv", encoding="utf-8")


DiagCN.drop(['Dname(ENG)'], axis=1, inplace=True)
DiagCN.rename(columns={'Dcode' : 'D_Code', 'Dname(KOR)' :'D_Name'}, inplace=True)
DiagCN=DiagCN.groupby(by='D_Code', as_index=False).agg({'D_Name': lambda a: a.value_counts().index[0]})

SurCN = SurCN.rename(columns={'Scode':'Sur_Code', 'Sname':'Sur_Name'})
SurCN=SurCN.groupby(by='Sur_Code', as_index=False).agg({'Sur_Name': lambda a: a.value_counts().index[0]})

DRGCNmain5['DRGNO'] = DRGCNmain5['DRGNO'].astype(str).str[0:4]
DRGCNmain5['DRGname'] = DRGCNmain5['DRGname'].str.split(', 연령').str[0]
DRGCNmain5=DRGCNmain5.drop_duplicates(['DRGNO'], keep='first')

DRGCNsub4=DRGCNsub4.drop_duplicates(['DRGNO'], keep='first')

## unite columns
SNUCHIn.drop(columns=['Birth'], inplace=True)
# KNUCHIn.drop(columns=['Birth'], inplace=True)
# SACHIn.drop(columns=['Birth'], inplace=True)
# SCHIn.drop(columns=['Birth'], inplace=True)
PNUCHIn.drop(columns=['Birth'], inplace=True)
# JNNUCHIn.drop(columns=['Birth'], inplace=True)
JBNUCHIn.drop(columns=['Birth'], inplace=True)
KBNUCHIn.drop(columns=['Birth'], inplace=True)


## DRGNO 4digit
SNUCHIn['DRGNO'] = SNUCHIn['DRGNO'].astype(str).str[0:4]
#KNUCHIn['DRGNO'] = KNUCHIn['DRGNO'].astype(str).str[0:4]
#SACHIn['DRGNO'] = SACHIn['DRGNO'].astype(str).str[0:4]
#SCHIn['DRGNO'] = SCHIn['DRGNO'].astype(str).str[0:4]
PNUCHIn['DRGNO'] = PNUCHIn['DRGNO'].astype(str).str[0:4]
JNNUCHIn['DRGNO'] = JNNUCHIn['DRGNO'].astype(str).str[0:4]
JBNUCHIn['DRGNO'] = JBNUCHIn['DRGNO'].astype(str).str[0:4]
KBNUCHIn['DRGNO'] = KBNUCHIn['DRGNO'].astype(str).str[0:4]

print(SNUCHIn.columns)
#print(KNUCHIn.columns)
#print(SACHIn.columns)
#print(SCHIn.columns)
print(PNUCHIn.columns)
print(JNNUCHIn.columns)
print(JBNUCHIn.columns)
print(KBNUCHIn.columns)


SubIns_Cri = pd.read_csv("/Users/peterpark/Desktop/DATA_ANALYSIS/Research/Dev4PubChildCenter/RawData/2018-2019SubIns_Cri.csv", encoding="utf-8", low_memory=False)

AllIn = pd.concat([SNUCHIn,
#                   KNUCHIn,
#                   SACHIn,
#                   SCHIn,
                   PNUCHIn,
                   JNNUCHIn,
                   JBNUCHIn,
                   KBNUCHIn], ignore_index=True)

print(AllIn.info())
print(AllIn.columns)

AllIn['PT_No'] = AllIn['PT_No'].astype(str)
AllIn['PT_No'] = AllIn[['InstName', 'PT_No']].apply(lambda x: ''.join(x), axis=1)
del AllIn['InstName']

AllIn.to_csv('./All/AllInP_R4A.csv', index=False)



############################# Analyzing #############################


## RAW DATA based on DRGNO by the order of payment: append rows
AllIn4SevPbase = AllIn.copy()
#AllIn4SevPbase.DRGNO = AllIn4SevPbase.DRGNO.str.split('/')
#AllIn4SevPbase.Severity = AllIn4SevPbase.Severity.str.split('/')
#AllIn4SevPbase = AllIn4SevPbase.apply(pd.Series.explode).reset_index(drop=True)
print(AllIn4SevPbase.info())
print(AllIn4SevPbase.columns)


DRGCNmain5.rename(columns={DRGCNmain5.columns[0] : 'DRGNO01', DRGCNmain5.columns[1] : 'DRGName01'}, inplace=True)
AllIn4SevPbase['DRGNO01'] = AllIn4SevPbase['DRGNO'].astype(str).str[0:4]
AllIn4SevPbase['DRGNO01'].fillna('NoDRG', inplace=True)

DRGCNsub4.rename(columns={DRGCNsub4.columns[0] : 'DRGNO02', DRGCNsub4.columns[1] : 'DRGName02'}, inplace=True)
AllIn4SevPbase['DRGNO02'] = AllIn4SevPbase['DRGNO'].astype(str).str[0:4]
AllIn4SevPbase['DRGNO02'].fillna('NoDRG', inplace=True)

AllIn4SevPbase = AllIn4SevPbase.merge(DRGCNmain5, on='DRGNO01', how='left').merge(DRGCNsub4, on='DRGNO02', how='left')
AllIn4SevPbase.reset_index(drop=True, inplace=True)

AllIn4SevPbase.loc[(AllIn4SevPbase['DRGName01'].isna() &  AllIn4SevPbase['DRGName02'].notna()), 'DRGName01'] = AllIn4SevPbase['DRGName02']
AllIn4SevPbase.rename(columns={'DRGName01' : 'DRGName'}, inplace=True)
del AllIn4SevPbase['DRGNO01']; del AllIn4SevPbase['DRGNO02']; del AllIn4SevPbase['DRGName02']
AllIn4SevPbase.DRGName.fillna('NoDRG', inplace=True)

AllIn4SevPbase.loc[(AllIn4SevPbase['DRGName']  == 'NoDRG'), 'DRGNO'] = 'NoDRG'
AllIn4SevPbase = AllIn4SevPbase[AllIn4SevPbase.D_Code != 'NoDiag']
print(AllIn4SevPbase.info())

AllIn4SevPbase['TotalPay'] = AllIn4SevPbase['Pay_InsSelf'] + AllIn4SevPbase['Pay_InsCorp'] + AllIn4SevPbase['Pay_NoIns'] + AllIn4SevPbase['Pay_Sel']

bpay=AllIn4SevPbase.groupby(by='DRGNO', as_index=False).agg({'TotalPay': pd.Series.mean})
bpay = bpay[bpay.DRGNO != 'NoDRG']
bpay.sort_values(by='TotalPay', ascending=False, inplace=True)
bpay.reset_index(drop=True, inplace=True)
bpay.index += 1
bpay = bpay.rename_axis('Rank').reset_index()
bpay['TotalPay'] = bpay['TotalPay'].astype(int)
print (bpay)

bdname=AllIn4SevPbase.groupby(by='DRGNO', as_index=False).agg({'DRGName': lambda a: a.value_counts().index[0]})
bdname = bdname[bdname.DRGNO != 'NoDRG']
bdname.reset_index(drop=True, inplace=True)
#print (bdname)


AllIn4SevPbase_tobsev = AllIn4SevPbase.copy()
AllIn4SevPbase_tobsev['Severity'] = AllIn4SevPbase_tobsev['Severity'].map({'Severe' : 4, 'Normal' : 3, 'Simple' : 2, 'SortError' : 1, 'NoDRG' : 0})

AllIn4SevPbase_tobsevsub = AllIn4SevPbase_tobsev.copy()
AllIn4SevPbase_tobsev = AllIn4SevPbase_tobsev.groupby(['DRGNO'], as_index= False)['Severity'].agg(lambda x: x.max())
AllIn4SevPbase_tobsev= AllIn4SevPbase_tobsev.merge(AllIn4SevPbase_tobsevsub, on=['DRGNO', 'Severity'], how='left')
AllIn4SevPbase_tobsev.drop_duplicates(subset =['PT_No', 'In_Date', 'Severity'], inplace = True)
AllIn4SevPbase_tobsev.reset_index(drop=True, inplace=True)

AllIn4SevPbase_tobsev['Severity'] = AllIn4SevPbase_tobsev['Severity'].map({4:'Severe', 3:'Normal', 2:'Simple', 1:'SortError', 0:'NoDRG'})

bsev=AllIn4SevPbase_tobsev.groupby(by='DRGNO', as_index=False).agg({'Severity': pd.Series.unique})
bsev.reset_index(drop=True, inplace=True)
print (bsev)


## 01. BASIC INFORMATION by Severity Payment
pnum=AllIn4SevPbase.groupby(by='DRGNO', as_index=False).agg({'PT_No': pd.Series.nunique})
pnum.rename(columns={'PT_No':'P_Num'}, inplace=True)
pnum['P_Num'] = pnum['P_Num'].astype(int)
pnum.reset_index(drop=True, inplace=True)
print (pnum)

avage=AllIn4SevPbase.groupby(by='DRGNO', as_index=False).agg({'Age': pd.Series.mean})
avage.rename(columns={'Age':'AvAge'}, inplace=True)
avage.AvAge = avage.AvAge.round(1)
avage.reset_index(drop=True, inplace=True)
print (avage)

avinprd=AllIn4SevPbase.groupby(by='DRGNO', as_index=False).agg({'In_Prd': pd.Series.mean})
avinprd.rename(columns={'In_Prd':'AvInPrd'}, inplace=True)
avinprd.AvInPrd = avinprd.AvInPrd.round(1)
avinprd.reset_index(drop=True, inplace=True)
print (avinprd)

AllIn4SevPbase_Ins = AllIn4SevPbase.copy()
AllIn4SevPbase_Ins['IndPaidExp'] = AllIn4SevPbase_Ins['Pay_InsSelf'] + AllIn4SevPbase_Ins['Pay_NoIns'] + AllIn4SevPbase_Ins['Pay_Sel']
inspaid = AllIn4SevPbase_Ins.groupby(by='DRGNO', as_index=False).agg({'IndPaidExp' : pd.Series.mean, 'Pay_InsCorp' : pd.Series.mean})
inspaid.rename(columns={'Pay_InsCorp':'CorpPaidExp'}, inplace=True)
inspaid['IndPaidExp'] = inspaid['IndPaidExp'].astype(int)
inspaid['CorpPaidExp'] = inspaid['CorpPaidExp'].astype(int)
inspaid['IndPaidExp'] = inspaid.apply(lambda x: "{:,}".format(x['IndPaidExp']), axis=1)
inspaid['CorpPaidExp'] = inspaid.apply(lambda x: "{:,}".format(x['CorpPaidExp']), axis=1)
inspaid.reset_index(drop=True, inplace=True)
print (inspaid)


AllInSevP_base = bpay.merge(bdname,on='DRGNO',how='left').merge(bsev,on='DRGNO',how='left').merge(pnum,on='DRGNO',how='left').merge(avage,on='DRGNO',how='left').merge(avinprd,on='DRGNO',how='left').merge(inspaid,on='DRGNO',how='left')
AllInSevP_base = AllInSevP_base[['Rank', 'DRGNO', 'DRGName', 'Severity', 'TotalPay', 'IndPaidExp', 'CorpPaidExp', 'P_Num', 'AvAge', 'AvInPrd']]
AllInSevP_base.reset_index(drop=True, inplace=True)
print(AllInSevP_base)
print(AllInSevP_base.columns)


AllInSevP_base.Severity = AllInSevP_base.Severity.map({'Severe' : '전문', 'Normal' : '일반', 'Simple' : '단순', 'SortError' : '분류오류'})

AllInSevP_baseSev = AllInSevP_base[AllInSevP_base.Severity == '전문']
AllInSevP_baseSev = AllInSevP_baseSev.sort_values(by ='TotalPay', ascending = 0)
AllInSevP_baseSev.drop(['Rank'], axis=1, inplace=True)
AllInSevP_baseSev.reset_index(drop=True, inplace=True)
AllInSevP_baseSev.index += 1
AllInSevP_baseSev['TotalPay'] = AllInSevP_baseSev.apply(lambda x: "{:,}".format(x['TotalPay']), axis=1)
AllInSevP_baseSev = AllInSevP_baseSev.rename_axis('Rank').reset_index()


AllInSevP_baseNorm = AllInSevP_base[AllInSevP_base.Severity == '일반']
AllInSevP_baseNorm = AllInSevP_baseNorm.sort_values(by ='TotalPay', ascending = 0)
AllInSevP_baseNorm.drop(['Rank'], axis=1, inplace=True)
AllInSevP_baseNorm.reset_index(drop=True, inplace=True)
AllInSevP_baseNorm.index += 1
AllInSevP_baseNorm['TotalPay'] = AllInSevP_baseNorm.apply(lambda x: "{:,}".format(x['TotalPay']), axis=1)
AllInSevP_baseNorm = AllInSevP_baseNorm.rename_axis('Rank').reset_index()


AllInSevP_baseSimp = AllInSevP_base[AllInSevP_base.Severity == '단순']
AllInSevP_baseSimp = AllInSevP_baseSimp.sort_values(by ='TotalPay', ascending = 0)
AllInSevP_baseSimp.drop(['Rank'], axis=1, inplace=True)
AllInSevP_baseSimp.reset_index(drop=True, inplace=True)
AllInSevP_baseSimp.index += 1
AllInSevP_baseSimp['TotalPay'] = AllInSevP_baseSimp.apply(lambda x: "{:,}".format(x['TotalPay']), axis=1)
AllInSevP_baseSimp = AllInSevP_baseSimp.rename_axis('Rank').reset_index()


AllInSevP_base['TotalPay'] = AllInSevP_base.apply(lambda x: "{:,}".format(x['TotalPay']), axis=1)


AllInSevP_base = AllInSevP_base.rename(columns={'Rank':'순위', 'DRGNO':'DRG번호', 'DRGName':'DRG명', 'Severity':'중증도', 'TotalPay':'총부담금(원)',
                                                    'IndPaidExp':'본인부담금(원)', 'CorpPaidExp':'공단부담금(원)', 'P_Num':'환자수(명)', 'AvAge':'평균연령(세)', 'AvInPrd':'평균재원기간(일)'})
AllInSevP_baseSev = AllInSevP_baseSev.rename(columns={'Rank':'순위', 'DRGNO':'DRG번호', 'DRGName':'DRG명', 'Severity':'중증도', 'TotalPay':'총부담금(원)',
                                                          'IndPaidExp':'본인부담금(원)', 'CorpPaidExp':'공단부담금(원)', 'P_Num':'환자수(명)', 'AvAge':'평균연령(세)', 'AvInPrd':'평균재원기간(일)'})
AllInSevP_baseNorm = AllInSevP_baseNorm.rename(columns={'Rank':'순위', 'DRGNO':'DRG번호', 'DRGName':'DRG명', 'Severity':'중증도', 'TotalPay':'총부담금(원)',
                                                            'IndPaidExp':'본인부담금(원)', 'CorpPaidExp':'공단부담금(원)', 'P_Num':'환자수(명)', 'AvAge':'평균연령(세)', 'AvInPrd':'평균재원기간(일)'})
AllInSevP_baseSimp = AllInSevP_baseSimp.rename(columns={'Rank':'순위', 'DRGNO':'DRG번호', 'DRGName':'DRG명', 'Severity':'중증도', 'TotalPay':'총부담금(원)',
                                                            'IndPaidExp':'본인부담금(원)', 'CorpPaidExp':'공단부담금(원)', 'P_Num':'환자수(명)', 'AvAge':'평균연령(세)', 'AvInPrd':'평균재원기간(일)'})

print(AllInSevP_base.columns)

AllInSevP_base50 = AllInSevP_base.loc[0:49, :]
AllInSevP_base50Sev = AllInSevP_baseSev.loc[0:49, :]
AllInSevP_base50Norm = AllInSevP_baseNorm.loc[0:49, :]
AllInSevP_base50Simp = AllInSevP_baseSimp.loc[0:49, :]

'''
AllInSevP_base.Severity = AllInSevP_base.Severity.map({'Severe' : '전문', 'Normal' : '일반', 'Simple' : '단순', 'SortError' : '분류오류'})


AllInSevP_baseSev = AllInSevP_base[AllInSevP_base.Severity == 'Severe']
AllInSevP_baseSev = AllInSevP_baseSev.sort_values(by ='TotalPay', ascending = 0)
AllInSevP_baseSev.drop(['Rank'], axis=1, inplace=True)
AllInSevP_baseSev.reset_index(drop=True, inplace=True)
AllInSevP_baseSev.index += 1
AllInSevP_baseSev['TotalPay'] = AllInSevP_baseSev.apply(lambda x: "{:,}".format(x['TotalPay']), axis=1)
AllInSevP_baseSev = AllInSevP_baseSev.rename_axis('Rank').reset_index()


AllInSevP_baseNorm = AllInSevP_base[AllInSevP_base.Severity == 'Normal']
AllInSevP_baseNorm = AllInSevP_baseNorm.sort_values(by ='TotalPay', ascending = 0)
AllInSevP_baseNorm.drop(['Rank'], axis=1, inplace=True)
AllInSevP_baseNorm.reset_index(drop=True, inplace=True)
AllInSevP_baseNorm.index += 1
AllInSevP_baseNorm['TotalPay'] = AllInSevP_baseNorm.apply(lambda x: "{:,}".format(x['TotalPay']), axis=1)
AllInSevP_baseNorm = AllInSevP_baseNorm.rename_axis('Rank').reset_index()


AllInSevP_baseSimp = AllInSevP_base[AllInSevP_base.Severity == 'Simple']
AllInSevP_baseSimp = AllInSevP_baseSimp.sort_values(by ='TotalPay', ascending = 0)
AllInSevP_baseSimp.drop(['Rank'], axis=1, inplace=True)
AllInSevP_baseSimp.reset_index(drop=True, inplace=True)
AllInSevP_baseSimp.index += 1
AllInSevP_baseSimp['TotalPay'] = AllInSevP_baseSimp.apply(lambda x: "{:,}".format(x['TotalPay']), axis=1)
AllInSevP_baseSimp = AllInSevP_baseSimp.rename_axis('Rank').reset_index()


AllInSevP_base['TotalPay'] = AllInSevP_base.apply(lambda x: "{:,}".format(x['TotalPay']), axis=1)


AllInSevP_base = AllInSevP_base.rename(columns={'Rank':'순위', 'DRGNO':'DRG번호', 'DRGName':'DRG명', 'Severity':'중증도', 'TotalPay':'총부담금(원)',
                                                    'IndPaidExp':'본인부담금(원)', 'CorpPaidExp':'공단부담금(원)', 'P_Num':'환자수(명)', 'AvAge':'평균연령(세)', 'AvInPrd':'평균재원기간(일)'})
AllInSevP_baseSev = AllInSevP_baseSev.rename(columns={'Rank':'순위', 'DRGNO':'DRG번호', 'DRGName':'DRG명', 'Severity':'중증도', 'TotalPay':'총부담금(원)',
                                                          'IndPaidExp':'본인부담금(원)', 'CorpPaidExp':'공단부담금(원)', 'P_Num':'환자수(명)', 'AvAge':'평균연령(세)', 'AvInPrd':'평균재원기간(일)'})
AllInSevP_baseNorm = AllInSevP_baseNorm.rename(columns={'Rank':'순위', 'DRGNO':'DRG번호', 'DRGName':'DRG명', 'Severity':'중증도', 'TotalPay':'총부담금(원)',
                                                            'IndPaidExp':'본인부담금(원)', 'CorpPaidExp':'공단부담금(원)', 'P_Num':'환자수(명)', 'AvAge':'평균연령(세)', 'AvInPrd':'평균재원기간(일)'})
AllInSevP_baseSimp = AllInSevP_baseSimp.rename(columns={'Rank':'순위', 'DRGNO':'DRG번호', 'DRGName':'DRG명', 'Severity':'중증도', 'TotalPay':'총부담금(원)',
                                                            'IndPaidExp':'본인부담금(원)', 'CorpPaidExp':'공단부담금(원)', 'P_Num':'환자수(명)', 'AvAge':'평균연령(세)', 'AvInPrd':'평균재원기간(일)'})

print(AllInSevP_base.columns)


AllInSevP_base50Sev = AllInSevP_baseSev.loc[0:49, :]
AllInSevP_base50Norm = AllInSevP_baseNorm.loc[0:49, :]
AllInSevP_base50Simp = AllInSevP_baseSimp.loc[0:49, :]
'''
#AllInSevP_base50 = AllInSevP_base.loc[0:49, :]

'''
# No Executuon
## 02. DIAGNOSIS & SURGERY INFORMATION by Severity Payment
AllIn4SevPbase_Diag = AllIn4SevPbase.copy()
AllIn4SevPbase_Diag.D_Code = AllIn4SevPbase_Diag.D_Code.str.split('/')
AllIn4SevPbase_Diag.D_Name = AllIn4SevPbase_Diag.D_Name.str.split('/')
AllIn4SevPbase_Diag = AllIn4SevPbase_Diag.apply(pd.Series.explode).reset_index(drop=True)
AllIn4SevPbase_Diag = AllIn4SevPbase_Diag.drop_duplicates(['PT_No', 'In_Date', 'DRGNO', 'Severity', 'D_Code', 'D_Name'], keep='first') # Need D_Date after for proper drop
AllIn4SevPbase_Diag = AllIn4SevPbase_Diag[AllIn4SevPbase_Diag.D_Code != 'NoDiag']

##### another way to selcet most frequent value using groupby(Not Accurate in Certain Condition)#####
#dcodename = AllIn4SevPbase_Diag.groupby(by='DRGNO', as_index=False).agg({'D_Code': lambda a: ' / '.join(pd.Series.mode(a)), 'D_Name': lambda b: ' / '.join(pd.Series.mode(b))})


dcode = AllIn4SevPbase_Diag.groupby(by='DRGNO', as_index=False)['D_Code'].agg(lambda x:x.value_counts().index[0])
dname = AllIn4SevPbase_Diag.groupby(by='D_Code', as_index=False).agg({'D_Name': lambda a: '/'.join(pd.unique(a))})
dcode.reset_index(drop=True, inplace=True)
dname.reset_index(drop=True, inplace=True)
dcodename = dcode.merge(dname, on='D_Code', how='left')
dcodename.reset_index(drop=True, inplace=True)
dfreq = AllIn4SevPbase_Diag.groupby(by='DRGNO', as_index=False)['D_Code'].agg(lambda x: x.value_counts().head(1))
dfreq.rename(columns={'D_Code':'Dfreq'}, inplace=True)
dfreq['Dfreq'] = dfreq['Dfreq'].astype('int64')
dfreq.reset_index(drop=True, inplace=True)
dtot = AllIn4SevPbase_Diag.groupby(by='DRGNO', as_index=False)['D_Code'].agg(lambda x: x.count())
dtot.rename(columns={'D_Code':'Dtot'}, inplace=True)
dtot['Dtot'] = dtot['Dtot'].astype('int64')
dtot.reset_index(drop=True, inplace=True)
dcodenamefreq = dcodename.merge(dfreq,on='DRGNO',how='left').merge(dtot,on='DRGNO',how='left')
dcodenamefreq.rename(columns={'D_Code':'Dcode', 'D_Name':'Dname'}, inplace=True)
dcodenamefreq.reset_index(drop=True, inplace=True)
dcodenamefreq['Dratio'] = (dcodenamefreq.Dfreq / dcodenamefreq.Dtot) * 100
dcodenamefreq.Dratio = dcodenamefreq.Dratio.round(2)
dcodenamefreq['Dratio'] = dcodenamefreq['Dratio'].astype(str) + ' %'
with pd.option_context('display.max_columns', None):
    print (dcodenamefreq)


AllIn4SevPbase_Sur = AllIn4SevPbase.copy()
AllIn4SevPbase_Sur.Sur_Date = AllIn4SevPbase_Sur.Sur_Date.str.split('/')
AllIn4SevPbase_Sur.Sur_Code = AllIn4SevPbase_Sur.Sur_Code.str.split('/')
AllIn4SevPbase_Sur.Sur_Name = AllIn4SevPbase_Sur.Sur_Name.str.split('/')
AllIn4SevPbase_Sur = AllIn4SevPbase_Sur.apply(pd.Series.explode).reset_index(drop=True)
AllIn4SevPbase_Sur = AllIn4SevPbase_Sur.drop_duplicates(['PT_No', 'In_Date', 'DRGNO', 'Severity', 'Sur_Date', 'Sur_Code', 'Sur_Name'], keep='first')
AllIn4SevPbase_OnlySur = AllIn4SevPbase_Sur[AllIn4SevPbase_Sur.Sur_Code != 'NoSur']
AllIn4SevPbase_OnlySur = AllIn4SevPbase_OnlySur[AllIn4SevPbase_OnlySur.Sur_Code != 'GroupPay_All']

##### another way to selcet most frequent value using groupby(Not Accurate in Certain Condition)#####
#scodename = AllIn4SevPbase_Sur.groupby(by='DRGNO', as_index=False).agg({'Sur_Code': lambda a: ' / '.join(pd.Series.mode(a)), 'Sur_Name' : lambda b: ' / '.join(pd.Series.mode(b))})


scode = AllIn4SevPbase_OnlySur.groupby(by='DRGNO', as_index=False)['Sur_Code'].agg(lambda x : x.value_counts().index[0])
sname = AllIn4SevPbase_OnlySur.groupby(by='Sur_Code', as_index=False)['Sur_Name'].agg(lambda x : x.value_counts().index[0])
scode.reset_index(drop=True, inplace=True)
sname.reset_index(drop=True, inplace=True)
scodename = scode.merge(sname, on='Sur_Code', how='left')
scodename.reset_index(drop=True, inplace=True)
sfreq = AllIn4SevPbase_OnlySur.groupby(by='DRGNO', as_index=False)['Sur_Code'].agg(lambda x : x.value_counts().head(1))
sfreq.rename(columns={'Sur_Code':'Sfreq'}, inplace=True)
sfreq['Sfreq'] = sfreq['Sfreq'].astype('int64')
sfreq.reset_index(drop=True, inplace=True)
stot = AllIn4SevPbase_OnlySur.groupby(by='DRGNO', as_index=False)['Sur_Code'].agg(lambda x : x.count())
stot.rename(columns={'Sur_Code':'Stot'}, inplace=True)
stot['Stot'] = stot['Stot'].astype('int64')
stot.reset_index(drop=True, inplace=True)
scodenamefreq = scodename.merge(sfreq,on='DRGNO',how='left').merge(stot,on='DRGNO',how='left')
scodenamefreq.rename(columns={'Sur_Code':'Scode', 'Sur_Name':'Sname'}, inplace=True)
scodenamefreq.reset_index(drop=True, inplace=True)
scodenamefreq['Sratio'] = (scodenamefreq.Sfreq / scodenamefreq.Stot) * 100
scodenamefreq.Sratio = scodenamefreq.Sratio.round(2)
scodenamefreq['Sratio'] = scodenamefreq['Sratio'].astype(str) + ' %'
with pd.option_context('display.max_columns', None):
    print (scodenamefreq)


AllInSevP_diagsur = bpay.merge(bsev,on='DRGNO',how='left').merge(dcodenamefreq,on='DRGNO',how='left').merge(scodenamefreq,on='DRGNO',how='left')
AllInSevP_diagsur.drop(['TotalPay' ,'Dtot', 'Stot'], axis=1, inplace=True)
AllInSevP_diagsur.reset_index(drop=True, inplace=True)
AllInSevP_diagsur['Dcode'].fillna('NoDiag', inplace=True)
AllInSevP_diagsur['Dname'].fillna('NoDiag', inplace=True)
AllInSevP_diagsur['Dfreq'].fillna(0, inplace=True)
AllInSevP_diagsur['Dratio'].fillna('0.0 %', inplace=True)
AllInSevP_diagsur['Scode'].fillna('NoSur', inplace=True)
AllInSevP_diagsur['Sname'].fillna('NoSur', inplace=True)
AllInSevP_diagsur['Sfreq'].fillna(0, inplace=True)
AllInSevP_diagsur['Sratio'].fillna('0.0 %', inplace=True)

with pd.option_context('display.max_columns', None):
    print (AllInSevP_diagsur)

print(AllInSevP_diagsur.columns)

AllInSevP_diagsur50 = AllInSevP_diagsur.loc[0:49, :]
'''

AllInSevP_base50.to_csv('./All/All 입원환자 중증도 총부담금별 순위.csv', encoding='cp949', index=False)
AllInSevP_base50Sev.to_csv('./All/All 입원환자 중증도 총부담금별 순위(전문상위50).csv', encoding='cp949', index=False)
AllInSevP_base50Norm.to_csv('./All/All 입원환자 중증도 총부담금별 순위(일반상위50).csv', encoding='cp949', index=False)
AllInSevP_base50Simp.to_csv('./All/All 입원환자 중증도 총부담금별 순위(단순상위50).csv', encoding='cp949', index=False)
#AllInSevP_diagsur50.to_csv('./All/AllInSevP_diagsur50.csv', index=False)

AllInSevP_base.to_csv('./All/(원)All 입원환자 중증도 총부담금별 순위.csv', encoding='cp949', index=False)
AllInSevP_baseSev.to_csv('./All/(원)All 입원환자 중증도 총부담금별 순위(전문상위50).csv', encoding='cp949', index=False)
AllInSevP_baseNorm.to_csv('./All/(원)All 입원환자 중증도 총부담금별 순위(일반상위50).csv', encoding='cp949', index=False)
AllInSevP_baseSimp.to_csv('./All/(원)All 입원환자 중증도 총부담금별 순위(단순상위50).csv', encoding='cp949', index=False)

#AllInSevP_inst = pd.DataFrame(columns=['Rank' ,'DRGNO', 'Severity', 'A_Ratio', 'B_Ratio', 'C_Ratio', 'D_Ratio', 'E_Ratio', 'F_Ratio', 'G_Ratio', 'H_Ratio'])
#AllInSevP_regn = pd.DataFrame(columns=['Rank' ,'DRGNO', 'Severity', 'fSeoul', 'fBusan', 'fDaegu', 'fGwangju', 'fDaejeon', 'fIncheon', 'fJeju', 'fSejong', 'fJeonnam', 'fJeonbuk',
#                                      'fGyeongnam', 'fGyeongbuk', 'fChungnam', 'fChungbuk', 'fGangwon', 'fGyeonggi'])
